requirements.txt
black
isort
lapx
matplotlib
numpy
opencv-python
pyYAML
pytest
pytest-cov
python-dotenv
ray[tune]
numpy
requests
scipy
setuptools
tqdm
ultralytics
wandb
